refactor: 이메일 인증 로직 통합

+++ giveLunch/src/main/java/main/givelunch/properties/SecurityProperties.java
@@ -21,7 +21,7 @@ public record SecurityProperties(
                 int maxFailedAttempts,
 
                 @DefaultValue("15")
-                long lockMinutes
+                int lockMinutes
         ) {
         }
 }
